17 user testsdata valdiation exceptions

+++ src/main/java/com/example/waytogo/user/service/impl/UserServiceJPA.java
@@ -14,6 +14,7 @@ import org.springframework.util.StringUtils;
 
 import java.util.Optional;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicReference;
 
 @RequiredArgsConstructor
 @Service
@@ -47,19 +48,24 @@ public class UserServiceJPA implements UserService {
 
     @Override
     public UserDTO updateUserById(UUID userId, UserDTO userDTO) {
-        userDTO.setUserId(userId);
+        userDTO.setId(userId);
         return userMapper.userToUserDto(userRepository.save(userMapper.userDtoToUser(userDTO)));
     }
 
     @Override
-    public void deleteUserById(UUID userId) {
-        userRepository.deleteById(userId);
+    public boolean deleteUserById(UUID userId) {
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public void patchUserById(UUID userId, UserDTO userDTO) {
+    public Optional<UserDTO> patchUserById(UUID userId, UserDTO userDTO) {
+        AtomicReference<Optional<UserDTO>> atomicReference = new AtomicReference<>();
 
-        userRepository.findById(userId).ifPresent(foundUser -> {
+        userRepository.findById(userId).ifPresentOrElse(foundUser -> {
             if (StringUtils.hasText(userDTO.getPassword())) {
                 foundUser.setPassword(userDTO.getPassword());
             }
@@ -69,8 +75,12 @@ public class UserServiceJPA implements UserService {
             if (StringUtils.hasText(userDTO.getUsername())) {
                 foundUser.setUsername(userDTO.getUsername());
             }
-            userRepository.save(foundUser);
+            atomicReference.set(Optional.of(userMapper.userToUserDto(userRepository.save(foundUser))));
+        }, () -> {
+            atomicReference.set(Optional.empty());
         });
+
+        return atomicReference.get();
     }
 
     private PageRequest buildPageRequest(Integer pageNumber, Integer pageSize) {
